--- app/new-simple/page.tsx
@@ -1,39 +1,0 @@
-"use client";
-
-import { useChat } from "ai/react";
-
-export default function Chat() {
-  const { messages, input, handleInputChange, handleSubmit } = useChat({
-    onError(error) {
-      console.log("error", error);
-    },
-    onFinish(message) {
-      console.log("onFinish", message);
-      window.alert("onFinish");
-    },
-    onResponse(data) {
-      console.log("response", data);
-    },
-  });
-
-  return (
-    <div className="stretch mx-auto flex w-full max-w-md flex-col py-24">
-      {messages.length > 0
-        ? messages.map((m) => (
-            <div key={m.id} className="whitespace-pre-wrap">
-              {m.content}
-            </div>
-          ))
-        : null}
-
-      <form onSubmit={handleSubmit}>
-        <input
-          className="fixed bottom-0 mb-8 w-full max-w-md rounded border border-gray-300 p-2 shadow-xl"
-          value={input}
-          placeholder="Say something..."
-          onChange={handleInputChange}
-        />
-      </form>
-    </div>
-  );
-}
